Problem: The racket packages have circular dependencies
Solution: Break circular dependencies.

I'm breaking it in whatever direction is discovered last. We will see
if the packages work anyway, otherwise we're in a tricky situation.
gui, rackunit, other important packages depend on e.g. racket-doc, and
that one has circular dependencies.

The zip files are downloading now and package installs are running.
It's taking its while. For racket-doc and dependencies, the nix file
is 6874 lines and 604K.

That sounds a lot to me, but then hackage-packages.nix is hundreds of
thousands of lines in a 9 Mb file, so I guess it's fine.