WIP: raco install on deps probably does too much
Working on solution: Use config.rktd to make our current nix package
the installation scope of Racket, and refer to eh already-installed
racket packages in other nix packages by adding to the various search
paths.

This commit makes sure each nix package has a proper links.rktd file,
and that it can find those files of its dependencies, but it's not yet
working.

https://www.mail-archive.com/racket-users@googlegroups.com/msg37277.html
indicates that we may need to add pkg-search-dirs as well, for the
benefit of the pkg layer, which refers to the packages themselves by
name, not through the names created in the collections layer.