Problem: raco setup is causing trouble
During setup it tries to find, for certain packages, collects
directories that aren't there and never will be. I haven't
found a quick solution for this, or understood the real cause.

Solution: Don't do setup.

It's entirely possible that this will bite us later, but for the
packages coming from racket-lang, this is appropriate.

We are now able to build over 500 packages. When it comes to
scribble-doc, which has 551 transitive deps, we choke on command-line
limitations because of how we generate config.rktd.