Problem: Skipping circular dependencies breaks install
When we simply do not install a circular dependency, the dependant
fails to install due to the missing dependency.

Solution: Remove the dependencies inside the packages when installing.

We go inside the info.rkt and remove all references to the circular
dependency.

This works for the test set of racket-doc with transitive
dependencies.

drracket builds again, and this time it's not empty. It's not complete
though, that's for issue #23.

Closes #19