Problem: drracket doesn't build
There are some circular dependencies in there. They can be made to
work.

The top package is the one that was detected as a circular dependency
by the bottom package and was cut out of the circle. The top package
transitively depends on the rest of the circle, the others are cut off
at the bottom package.

Here's how we can make this work:

1. Don't run setup on any bottom package.

There are some true circular dependencies in there, that won't work
unless the dependency has had a proper setup and registered its
collections. They can be made to work by doing some extra work in the
top (not necessarily, but we were lucky enough that it was the top --
the circle was only two packages) package:

2a. Remove the bottom package from the config.rktd dependencies.
2b. Install both the top and bottom packages in the top installation
    scope.
2c. Run setup on both.

Solution: To get a starting point to work from, first make some ugly
special cases to solve these specific packages. Work on a generic
solution from here.

htdp-lib and deinprogramm-signature have a true circular dependency
that can be solved by (2) (verified with htdp-lib as top).

racket-doc and drracket have a circular dependency that works by just
doing (1) -- force both to not do any setup for now. The racket-doc
package has a whole drove of circular dependencies with other -doc
packages.

For discussion of next steps, see issue #38.