Problem: We can't generate our own default.nix
We're not on the web, we're in the filesystem.
Racket package zips don't put the package in a subdirectory, it's right in the root, but unpackFile adds a subdirectory when it copies from the filesystem.

Solution: Support file system paths in derivations, detect the extra subdirectory level and compensate.

Now we have a Makefile:
 - Want to test? make test.
   This will verify that racket-doc builds,and it will show you if default.nix is generated differently.
 - Want to update default.nix according to your new code? Just make.
   You need to do this before pushing if you want the test to pass (once we add the .travis.yml to run make test on every PR).
 - Want a pkgs-all, a local copy of the release pkgs repo? make pkgs-all.
   You want this if you want to generate different nixes with `./racket2nix --catalog pkgs-all <the-package>`.
   You can do it without the --catalog parameter, but you'll download it from the network every time.