Problem: README is not valid asciidoc
I tried to generate the HTML for the README locally, and got:

    asciidoc: ERROR: README.adoc: line 4: only book doctypes can contain level 0 sections

I caused this when I made the update to README before, as I figured it doesn't seem
to make sense to not make use of all the levels. I stand corrected.

Solution: Add one level to all headlines.