Problem: `make` doesn't test that self-hosted racket2nix works
We would still catch it in Travis when we build racket-doc, but better
to catch it early.

Solution: Use an unpackaged racket2nix to create a default.nix, make a
new racket2nix from that, run again, compare outputs.

This replaces and improves the test in Makefile (you would have to run
the original `make` twice to get the equivalent), so we can lift that
out. If nix-build returns something, it's tested.