Problem: config.rktd is generated in a clumsy way
Dumping things into a text template is rigid and awkward. The template
interface is brittle to work with.

Solution: Provide better metadata in the derivation, generate the file
based on that metadata, with a racket script.

As a bonus, we further shrink the size of the nix expressions. Before:

    $ wc -l /nix/store/43gg9bmx28ay7rcw0y77gpimbfvqm0z0-racket-doc.nix
    2916 /nix/store/43gg9bmx28ay7rcw0y77gpimbfvqm0z0-racket-doc.nix

After:

    $ wc -l /nix/store/fd2a6mwv25mc9dxalagszalwmmki233h-racket-doc.nix
    1488 /nix/store/fd2a6mwv25mc9dxalagszalwmmki233h-racket-doc.nix