Problem: Storing and updating default.nix is an extra step
Solution: Make the self-hosted generated default.nix our implied default.nix

We lose the check that we haven't modified racket2nix in a way that
affects default.nix, but most of the time modifying the output is
exactly what we want to achieve, and we do check if we are still
self-hosting without any issues.

Without an old default.nix to verify against, test-racket2nix becomes
redundant, so building racket-doc becomes our complete test.nix.