Problem: raco setup tries to write to the racket directories
racket has internal assumptions that:
lib-dir means both where we can find launcher-sh and where we should write launchers.rktd
bin-dir means both where we can find racket and where we should write our launchers

Solution: Symlink racket in, clear it out after install, but leave bin because the launcher points to that racket.

This is not a complete solution -- the launcher doesn't put in the -G parameter necessary to find our collections.