Problem: reverse circular dependencies not removed
When we changed config.rktd in 318f237 one detail was lost: When the
top circular package wants to install a lower package within the top's
derivation, it needs to remove the lower package from its config.rktd
dependencies.

Otherwise 'raco pkg install' give us the error message that the
package is already installed in a larger scope.

Solution: Add the filtering mechanism on the nix side.

racket-doc doesn't build yet.