Problem: racket-doc deps don't build due to circularity
Solution: Add manual descriptions of the cycles.

This gives us a green build again. From here we can work on the real
solution. I have a plan.

We are still not trying to build racket-doc itself. I put hours into adding
manual cycles, and it led nowhere. We need to do this properly.