Problem: our Travis folds are too spammy
racket2nix now has dependencies, and the folds we have are too coarse.

Solution: Bring back the derivation folds filter, and apply it within
each major fold to produce sub-folds.

Travis doesn't actually have sub-folds, so let's emulate them.

The most interesting folds are the last of the
..-nix.drv..racket2nix and ..-nix.drv..racket2nix.full-racket folds,
because that's where the tests run. Maybe we should highlight that
somehow, but if the tests fail it becomes unfolded and obvious, so
it's not a priority for now.