Problem: pkgs path name is implicit
We are letting Nix figure out what to name a source derivation, and
that affects the path name in share/racket/pkgs once that source is
installed by raco.

Up until now this has happened to work, because all dependencies have
designated a subdirectory within the git repo, but
fractalide/fractalide uses the root, and a package within
cardano-wallet couldn't depend on it because it was looking for
'fractalide' and the pkgs directory only had
'fractalide-<short hash>'.

Solution: Name the git source derivation explicitly.

We are still relying on ?path= derivations and non-git source
derivations to accidentally have the right name. For everything
on pkgs.racket-lang.org this still happens to be true.