Problem: build-racket does not work with restrict-eval = true
```
$ nix-build --option restrict-eval true \
  -I . -I /nix/store/8wpfckjbdwhrfrfdra6gi2f69brdyz20-nixpkgs-18.09pre140705.090b7cc8f1b/nixpkgs \
  build-racket.nix --arg package ./nix
building '/nix/store/dh56a2qmy6x65pzscxn85dkr74pkbx9v-racket-package.nix.drv'...
installing
error: access to path '/nix/store/x8fbigk4ms45j7dpfcdi97x834mmynrd-racket-package.nix-src/nix' is forbidden in restricted mode
```

Solution: When a package is located inside the store, racket2nix
refers to it as a no-op fixed-output derivation instead of as a plain
path. Example:

```
fractalide/fractalide/pkgs $ nix-build --arg racket2nix 'import ../../racket2nix {}' -A fractalide.nix | xargs grep -m 1 -A 5 _fractalide
  _fractalide = mkRacketDerivation rec {
  name = "fractalide";
  src = fixedRacketSource {
    pathname = "/nix/store/glcimhq0swvnixz0p1nyksyfcp8xql5s-racket-package.nix-src/fractalide";
    sha256 = "4918fba598b3b18a028553121f275b9f1d4558cff4eb27cea2c66ffceb5dbd25";
  };
```

`pathname` is for human eyes only, the derivation does nothing with it
except use the basename as name.

The act of taking the hash of the directory makes it appear in the
store, that's why the derivation can be a no-op.

Verified to work with `racket2nix` and with `fractalide`.

Remove dependencies on nix-prefetch-git in our nixes. If you run
racket2nix through nix, you should use a preprocessed catalog instead.

racket2nix now needs nix whenever you pass it a directory rather than
a package name, because it needs to check if that directory is in the
store. There is a narrow usecase for using racket2nix without nix
installed, but I don't think it's worth it.

Closes #126