Problem: racket-build has unnecessary complexity
The `racket-package.nix.src` dance is for the purpose of making sure
the racket derivation depends on the subtree only, and not the whole
source repo. Now we have the fixed-output dance, which takes care of
that too.

Solution: Remove the `src` output from the `racket-package.nix`
derivation.

Add build-racket-packaged nix to the Travis tests, while also phasing
in use of nix-build-travis-fold.sh.

This nix-build-travis-fold.sh also uses restrict-eval, so that we
catch those issues before we reach Hydra.