Problem: jobs in release.nix are unordered
We make use of the ordering in travis-build.sh to make sure any
prerequisites get evaluated in a separate fold, so that once you get
to the meat of the build, you only see the most interesting
derivations clustered together.

Solution: Add explicit 'travisOrder' attribute.

This can also be used to exclude some of release.nix from being built
on Travis. Only the attributes named in travisOrder will be built.