Problem: dirname $PWD gives the wrong answer entirely
That should be `basename $PWD`. But that's wrong too, e.g. on
clacke/racket2nix-clacke the build directory is called
`racket2nix-clacke`.

Also the nixpkgs reference on NIX_PATH is not good enough for
restrict-eval, and our existing attempts to solve that cause new
problems when a `-I ''` occurred:

    error: getting status of '/home/travis/build/clacke/racket2nix-clacke/racket2nix/nixpkgs.nix'

Solution: Make the `racket2nix` nix path name explicit, and move it to
the caller.

Improve the mechanism for finding out where we are getting our nixpkgs from:

 - Only add a parameter if we actually found a link.
 - Add explicit /home/travis