Problem: No way to supply supplemental package paths
If you have several packages not in the catalog and want them to
depend on each other, you need to first generate a catalog for them.

Solution: Allow several package paths on the command line.

Package paths after the first one will be treated as supplemental and
used if the first one depends on them.

Closes #157