Problem: Some outdated workarounds are still around.
The racket2nix bootstrap refers to ../nix, then has to override it
when using the generated nix file. It can just refer to the nix
store alias instead to make it non-relative and remove the need
for an override.

Solution: Refer to $src instead of ../nix in the bootstraps.

Also make stage0 look more like a buildRacket-generated derivation.