Problem: No integration tests
We have unit tests for the catalog dependency resolver, but no quick
integration tests for the whole chain including Nix, and for specific
situations. In particular, we have no test for flattening a
dependency tree where a circular dependency is a dependency below
the root.

Solution: Create some thin, synthetic packages to test buildRacket on.

Unfortunately this cannot detect the current bug, because another bug
masks it -- the test should fail because we don't include the circular
dependencies of our circular dependency, but for some reason
c-depends-on-b is not removed from our transitive dependencies when it
is detected as circular, so the test passes (with the "should not have
depended on this package" warning message when not flattening).

But I want to get the test in anyway.