Problem: Our derivation names may collide with nixpkgs
Case in point: racket2nix itself is a package called 'nix', which
collides with the derivation of the nix package manager itself.

Solution: Prepend `$racket.name` to the package derivation name.

This has the added advantage of also distinguishing racket versions,
and even racket-full vs racket-minimal.

Also it solves another minor issue, which is that of distinguishing
the package derivation from the package source derivation.

Closes #127