Problem: No API for building catalogs
When you're in the situation where you have several packages not
published on pkgs.racket-lang, you have to create provisional
catalogs, patch them for internal vs external use, and run update
scripts outside Nix. See fractalide/cardano-wallet for the machinery
required.

Solution: Provide buildRacketCatalog, which takes several packages,
potentially with dependencies among themselves, and merges their
definitions with the default catalog to create a catalog that can
build either one of the packages.

Making one catalog per package and then merging them is a bit clunky,
and the real solution is to implement export of several packages in
racket2nix, but I'd like to kick that can down the road a bit, and
this workaround in Nix was quick and easy.

To be concrete, this is the kind of thing this change enables:

```
buildRacket {
  package = ./mypackage;
  catalog = buildRacketCatalog [
    ./mypackage
    ./myotherpackage
    (import ./mypkgs {}).theirpackage.src
  ];
}
```