Problem: Launcher creation is not pure
The launcher creation calls (make-relative-path-header), which peeks
at the filesystem for certain tools. How the output looks depends on
what it sees, and varies between sandboxed and non-sandboxed builds.

Solution: Set absolute-install? to #t in config.rktd.

This gets us a plain, hard-coded directory reference for launchers to
look for racket.

Closes #194

Makes any worries in #186 obsolete.