Problem: Attributes are prefixed with '_'
Now that we are referring to them with 'self.name' we can skip the workaround. Packages like 2
'2d-lib' can be safely addressed now that we can have quotes around the name.

Solution: Remove the '_' prefix, address attributes with a quoted string.

We are banking on nobody ever calling a racket package 'lib', 'extend'
or '__unfix__'. For the moment it holds.