Problem: Array output is not so useful
It would be more useful to output an attribute set of the packages.

But if you are going to access the packages by name anyway, why not
just output the whole (relevant subset of) the catalog?

Solution: When given > 1 package name, output attribute set.

The attribute set contains the packages named on the command line,
and all of their transitive dependencies.