Problem: No generator for extensions to racket-packages.nix
Solution: Implement --thin mode.

Now you can do this:

    $ nix-build -E "((import ./racket-packages.nix {}).extend ($(./racket2nix --thin ./nix))).nix"
    /nix/store/3qkwgcsqycza6igzj2cz7wa9gkb95j4s-racket-minimal-7.0-nix

This pushes some transitive dependency resolution to Nix, but it stops
when it encounters pre-calculated transitive dependencies. It won't
run into the circular dependencies in the catalog, and if it runs into
any circular dependencies outside the catalog, the user is to blame.

The main direct gain from this is that we can easily skip the catalog
generation step when building packages with dependencies not in the
racket catalog.

The generated expression looks like this:

    $ ./racket2nix --thin ./nix
    self: super: {
      nix = self.lib.mkRacketDerivation rec {
      pname = "nix";
      src = ./nix;
      racketThinBuildInputs = [ self."base" self."rackunit-lib" ];
      };

    }