Problem: integration-tests are simplistic
Solution: Convert package tree unit tests to integration tests.

Instead of the manually written packages, generate racket packages
with the full range of complications:
 - Long circles
 - Circles within circles
 - Independent circles depended on by other circles

This would have caught a thing or two that I encountered while
working on the new circle resolver (while the unit tests had not yet
been adapted).

For now, just verify that the packages build. The code in the packages
expresses the dependencies, so any missing dependency will cause a
compilation error.