Problem: Our racket-minimal-7.0 is a 6.12 disguised as 7.0
The basename of the source is racket-minimal-src.tgz, the version is
only reflected in the directory name.

We forgot to update the sha256 or the fixed-output derivation, so it
just used the old source under a new derivation name.

Solution: Update the sha256.

Nixpkgs has this issue too on nixpkgs-unstable. Fixed on master.