Problem: Our cycle detection is a dirty hack
We've been fudging and making a different package tree depending on
what you want to build, and hoping that you'd always depend on the
same package within a cycle, or that if some diamond dependency led in
through another part of the cycle, it's work itself out somehow. This
was good enough to build racket-doc, but not for generating a generic
racket-packages.nix.

The method is band-aid upon band-aid, with catalog mutation
everywhere for memoization and conflict resolution.

Solution: Implement cycle resolution in a few lines of neat datalog.