Problem: Doesn't put distributed-places in racket-doc cycle
Solution: Detect the important case when the current node is in
between the detected cycle bottom and cycle top.

Now builds racket-doc!

 - Keep track of all known cycles. When encountering one in the
   current node's dependencies, add us to the cycle if we belong.
   The previous implementation just avoided going down that
   branch, which might lead to packages being left out of cycles.

See https://libranet.de/display/0b6b25a8105b91ed740c2da648673068

This turned out to not be the actual reason distributed-places
was left out, but it might be relevant in some other case.