Problem: Flat builds may include the package twice
If the package is part of a cycle, the new cycle resolution makes the
cycle a concrete package that includes the original package. When the
package then includes all the inclusions of its transitive
dependencies, it adds itself to the list of extra sources, creating
a conflict.

Steps to reproduce:

    $ nix-build --argstr package racket-doc -A flat
    [ . . . ]
    mkdir: cannot create directory 'racket-doc': File exists
    builder for '/nix/store/jgvrik6zizy1kca7jqk9p0axinhr4sjq-racket-minimal-7.0-racket-doc.drv' failed with exit code 1
    error: build of '/nix/store/jgvrik6zizy1kca7jqk9p0axinhr4sjq-racket-minimal-7.0-racket-doc.drv' failed

Solution: When a flat package expands its inclusions, exclude itself.