Problem: Package with circular deps won't build in flat mode
When a package is built in flat mode, it retains the
circularBuildInputs attribute, which makes the derivation do nothing
as the package is considered a placeholder. This behavior is
incorrect for flat mode, as a flat derivation depends on no other
racket package.

Solution: In flat mode, circularBuildInputs is always the empty list.