Problem: Cannot use builtins.path on store paths
The hack to correct the directory name of the source only works when
running buildThinRacket on non-store paths.

In concrete terms: fractalide doesn't build with buildThinRacket.

Solution: Replace builtins.path with a fixed-output derivation.

This should duplicate the fixedRacketSource fixed-output derivation
that racket2nix creates anyway, so in the end no extra disk work is
actually done. Actually this probably saves work running racket2nix
when the same source is used in different contexts.