Problem: nix-instantiate cannot run in the sandbox
Our method for finding out whether racket2nix is running inside a Nix
derivation (i.e. called from buildRacket) ironically cannot run inside
a Nix derivation, if the sandbox is enabled.

    building '/nix/store/bksdhdyj4hw5ply332zypxmf8rxy9dik-racket-package.nix.drv'...
    error: creating directory '/nix/var': Permission denied
    builder for '/nix/store/bksdhdyj4hw5ply332zypxmf8rxy9dik-racket-package.nix.drv' failed with exit code 1

What it really wants to find out is whether its inputs are in the Nix
store and need to be wrapped in a fixed-output derivation to pass
strict evaluation.

Solution: For now, hard-code the store path as '/nix/store'.

Honestly, it not being '/nix/store' is an extreme corner case that may
not even work. It just didn't feel right to hard-code it.

Closes #226