Problem: Bumped live catalog brings more broken packages
Solution: Add more exclusions.

There is a new brand of breakage, where a dependency is a direct git
reference instead of a package name. We will have to check if this is
a new or previously unknown feature that we should support, or if it
is a mistake.

For now, remove the affected packages.