Problem: Error messages about SSL certificates
    openssl: x509-root-sources: cert sources do not exist: /no-cert-file.crt, /nix/store/nlb8hykwq7jc4vpnda39civ9cvrb7ymq-openssl-1.0.2p/etc/ssl/certs; override using SSL_CERT_FILE, SSL_CERT_DIR

It's only an annoyance -- we do not perform any network operations --
but it's the typical kind of warning message that might take someone
on a wild goose chase while troubleshooting.

Solution: Add `cacert` to the `buildInputs` of things running racket.

Allowing `cacert` to manipulate `SSL_CERT_FILE` is cleaner than
manipulating the variable directly. Modify the derivations that do.