Problem: bump-release-catalog uses system racket
We should use the racket the catalog is for.

Solution: Use nix-shell inside the script instead of as interpreter.

We cannot just skip using racket, as we did in #242 with the live
catalog, because the release catalog uses relative paths and we want
racket to sort that out for us.