Problem: We are parsing some github URLs wrong
One regex is ambiguous (for humans), and has had this error from the
very beginning, or at least for as long as racket-packages.nix has
been checked in. We were just (un)lucky enough to not happen to be
depending on any of the affected packages, so we never noticed.

Solution: Add unit test for the regression, fix regex.