Problem: No general mechanism for running package tests
Solution: Run raco test on all rkt files when doInstallCheck enabled.

Adapted from the fractalide-tests target in Fractalide, but:
 - Instead of creating a separate derivation it just creates an output.
 - The test output is less spammy.
 - Don't rely on parallel, xargs has parallelization too.
 - Don't just show output in build output, it may interleave. Output
   to per-file output files too.

Fractalide still can't make use of the complete generic mechanism,
because we have directories we want to skip. But the required work in
Fractalide gets reduced to enabling doInstallCheck and overriding
installCheckPhase, and that's an improvement too.