Problem: racket-packages.nix shortcut only in default.nix
This makes it a bit obscure to use from e.g. test.nix.

We want to use racket-packages.nix to get the full catalog
and not only the subset of packages in the transitive dependencies,
because the #251 installCheckPhase will want compiler-lib to be
available when building e.g. br-parser-tools-lib.

Also regenerating the Nix expression is unnecessary work.

Solution: Put the functionality in buildRacket instead.

Make default.nix an ever thinner wrapper around build(Thin)Racket.