Problem: doInstallCheck overrides: Not checking all checkable
Solution: Run update-install-check-overrides.

    real    400m11.041s
    user    8m35.924s
    sys     1m43.327s

These numbers on a reasonably non-primed system -- most packages were
built, not cached.  Parallelism would be a valuable addition. On the
other hand, we don't have to run this all that often.

Grandfathered in, would not get added by a fresh run (but would not be
removed by additional update-install-check-overrides, because it only adds to the list):

 - srfi-lib
 - typed-racket-lib

They don't have the typical rackunit output, specifically they lack
the "tests passed" line, but they do run tests.