Problem: pict-lib doesn't build with racket-full
It tries to run the tests, but the package is not there because it is
in the racket derivation.

There isn't much point in supporting racket-full, it was always just
'also there'. Actually, I thought I had already removed it. On
x86_64-darwin we don't even have one.

Solution: Remove racket-full jobs from release.nix.