Problem: Too much going on during eval
Because of how default.nix is set up, any use of buildRacket at all
involves bootstrapping and verifying racket2nix before you can get a
derivation or even a function.

The worst consequence of this right now is, when a build worker is
down, release.nix doesn't evaluate on hydra.

Solution: Make racket2nix from a static nix expression.

We check that the expression is up to date in release, just like with
racket-packages, and we do the full bootstrap and verification of
racket2nix in release only.