Problem: catalog, release-catalog: pretty-write not stable
On Linux, pretty-write sorts string keys for hashes, on macOS it
doesn't. This makes the live-catalog hash unreliable and it means that
catalog.rktd will be rendered differently depending on platform. And
if you render catalog.rktd on macOS, the diff will be uninformative.

Solution: Do our own sorting.

macOS Racket does sort symbol keys when pretty-writing, so the package
entries internally are fine.