Problem: test: runCommand does too much
In order to group several derivations under one job in Hydra, we are
wrapping them in a runCommand. However, runCommand runs stdenv's
setup.sh, and it does several things we don't need. In particular, it
sources any buildInput that is a file (see activatePackage).

It is by pure luck that catalog.rktd is treated as one big comment by
bash, because it is all on one line that happens to start with #hash.
If you pretty-print the catalog, the job depending on it breaks.

    $ nix-build test.nix -A light-tests
    these derivations will be built:
      /nix/store/iglbykb2ky13hq1g5kw40mkg237zmrnv-light-tests.drv
    building '/nix/store/iglbykb2ky13hq1g5kw40mkg237zmrnv-light-tests.drv'...
    /nix/store/1rvmv7gl87kc1b38ax4nq512xx5a6rg9-catalog.rktd: line 2: .: filename argument required
    .: usage: . filename [arguments]
    builder for '/nix/store/iglbykb2ky13hq1g5kw40mkg237zmrnv-light-tests.drv' failed with exit code 2

Solution: buildDrvs: The simplest possible derivation.