Problem: update-racket-packages not safe to run in parallel
This may sound like an exotic use case, but when iteratively fixing
things, running multiple update-racket-packages in parallel actually
makes sense.

You try to build something, it fails in multiple dependencies. You fix
one thing, start updating racket-packages to then run a test build of
that thing. While that is running, you fix the next thing, and want to
run tests for that. Currently you need to wait for the first update to
finish before starting the next, or they trip over each other's .new
files.

Solution: Use unique temporary files.

Now, when one update finishes, it can start running its build. If
while it's updating you make more changes, you can start an update
based on those, and then build based on those changes as soon
as the second update is done. The latest finished update wins.

Of course it is possible to confuse yourself about which changes are
in effect, but that's up to you. Generally this doesn't come up,
because there's a general direction of "forward", and if you get
some further updates than expected, it's not an issue.

    # make some changes then ...
    $ ./update-racket-packages.sh && nix-build --argstr package thing
    # while that is running, make more changes then run in parallel ...
    $ ./update-racket-packages.sh && nix-build --argstr package otherthing