Problem: catalog output still unstable
I ran the code on a third system (NixOS with racket 7.1) and there the
small internal hashes came out different. So there is no escape, we
have to sort all hashes.

Solution: Write our own pretty-writer.

I tried jacking into the existing writer system with
(port-print-handler) or (gen:custom-write) to try to reuse the
existing pretty-writer, but it's not trivially done. This is simpler.

This is not a generic pretty-printer. It is exactly as good as it
needs to be for our purposes. On a positive note, it produces more
compact and readable output for our particular data than
(pretty-write) does.