Problem: Unproductive caching and remoting
Simple tasks like copying a file from one place to another, or
processing some small text file, should not incur the overhead and
latency of talking to upstream caches or build workers. Just do it.

Solution: Set preferLocalBuild and allowSubstitutes where appropriate.

Applies to most runCommands, but not the ones in catalog.nix, which
are reasonably time-consuming.