Problem: test fixed-output srcs work by accident
This was made apparent by #271 / ff4dc0d and running tests on a clean
machine. Before that fix, it worked because the fixed-output
derivations were being downloaded from hydra.

This is one of those "but how could this have ever worked" things. Not
sure if something changed in Nix or racket2nix that made it not work.

Solution: Use direct derivations instead of generated fixed-output.

This means overriding the racket-packages output from buildRacket.
There is opportunity here for several usability fixes that would make
the fix less kludgy, and which would benefit real buildRacket users.

It should be possible to use buildRacketCatalog and buildRacket
together and it Just Works, or even better, not having to use
buildRacketCatalog.

For well-behaved packages there is buildThinRacket, which Just Works,
but integration-tests is about circular dependencies among the
packages supplied, and for that you currently need buildRacketCatalog
and the full buildRacket.