Problem: integration-tests broken
Since the 'fix' in 627a958b #273 the integration-test packages have
been confused about what to build, e.g. 'a' is part of a cycle, but
still tries to build from its own source.

    $ ./release.sh -A tests.light-tests
    [ . . . ]
      for package: ./a
      missing packages:
       base
       b
    [ . . . ]

Solution: Override with overrideRacketDerivation.

Fudging with overrideAttrs upset some internal assumption. There is
work to do here on the derivation hygiene, for sure.

The error went undetected because of a bug in ./release.sh
(the -A tests.light-tests is required for the test to run).
Separate PR incoming.