Problem: We have manually maintained compilation workarounds
Some packages cause the compiler to want to recompile dependencies, so
for those we have hijacked part of the circular-dependencies mechanism
to compile the dependency and the main package together, so that the
dependency is writable.

Some of this propagates. If e.g. racket-index has this issue
with scribble-lib, and you have this issue with racket-index, now you
need to add both packages to your writable dependencies.

Under 7.2 the manual effort is prohibitive, this occurs in too many
packages. Under 7.3 it gets even worse.

Solution: Detect when setup fails, retry with *all* deps writable.

Put the writable dependencies outside the derivation, so that
the issues doesn't propagate the same way (and also we have lighter
derivations, and no duplicates in the installed package set).

The `force-reverse-circular-build-inputs` hash of workarounds is now
empty!

 * Put racket-index to light-tests, as it is an indicator of this
   issue.