Problem: installCheck: Error output can be confusing
The bash process running raco test is named as the file it is testing,
so you can get a confusing error message along the lines of:

    .../info.rkt: foo: unbound variable

What would an info.rkt want with a variable? It's bash that is using
the process name to label its error messages.

Solution: Add some context to the process name.