Problem: bcrypt out and env have a circular dependency
The problem is that nixpkgs automatically sets `-rpath` pointers in
native libraries to `out`, but we keep our native libraries in `env`
together with the compiled racket libraries.

Our wrappers in `out` point to the libraries in `env`, so if the
libraries point back to `out` we get a circular dependency.

Solution: Rename the `env` output to `lib`.

nixpkgs multiple-outputs.sh prioritizes the `lib` output for rpaths over
`out` if it exists.

```
cycle detected in the references of '/nix/store/...-bcrypt-env' from '/nix/store/...-bcrypt'
error: build of '/nix/store/...-bcrypt.drv' failed
```