Problem: Explicit racketBuildInputs is redundant
It blows up the size of racket-packages.nix and makes it harder to
read and also harder to manipulate with overrides.

Solution: Retire racketBuildInputs as an explicit parameter.

Use racketThinBuildInputs everywhere.

 - Handle reified cycle packages in deps instead of in transitive
   deps.
 - Instead of each package resolving thin inputs for all its deps,
   rely on dependencies to resolve their dependencies.
 - Deduplicate the transitive dependencies on Nix side. This used to
   be done on the Racket side. Without it, building even racket2nix
   runs out of Nix heap.

There is refactoring to do and dead code to remove, but there will be
even more dead code when the cycle handling too is moved to Nix, so
just hold off on all that.